fix: move homepage to undergo resultlayout
